Rework stave drawing to honour width
The previous version had a hard-coded width for the stave lines. This
worked for single-column A4 portrait, but not anything else. In
particular it failed for the cheat sheet.

This 'fix' draws new stave lines after each note from the left hand end
of the line to just after the note. This means the stave is always
just the right length, at the expense of repeatedly redrawing the lines.
MacOS preview (at least) seems to find this confusing when drawing
a zoomed-out view though it seems to be OK if you zoom in or print.

This will have to do for now.