Implement a better fix for mandolin tab width
The previous fix for the width of the mandolin tablature was
unsatisfactory because the stave lines were repeatedly redrawn.
MacOS preview and Chrome (at least) seemed to find this confusing
when drawing a zoomed-out view.

This fix reverts all the previous changes and works instead
by reading the width from a PostScript variable which has a useful
default but which can be overridden from a .fmt file as needed.
The result is much tidier.