Rework dulcimer versions to omit tabs for tunes not in 'D'
The dulcimer tabs correctly represented all the melodies, but playing
tunes in keys other then 'D' would be unsatisfactory because the drone
strings would be out of tune.

There's no easy fix for this. Pro-tem, this fix removes the tabs from
the existing Dulcimer tunebook (just leaving the cords), and creates
a new version with both chords and tabs but just for tunes in 'D'.

To do this, sorter.py was enhanced with the ability to filter tunes
based on the K: header.