Re-work Dulcimer music again
The new plan is to have:

* Melody tabs (for noter-drone play) for instruments with melody
  strings tuned to particular notes such as 'D' (from DAD) and
  perhaps other notes like 'A' (from DAA)
* Chords (for chord play) for instruments tuned DAD and perhaps
  other tunings

This edit reworks the previous 'chords and TABs for DAD tuning
but only for tunes in D' music into 'just TABs for all tunes
assuming a D melody string'.

This edit also changes the string labelling to use the passed-in
'pitch' value for the melody string and to abel the base and
middle strings with '?' to emphasise that they may vary.