Fix key (K) header for tunes not in a major key
K was previously set by reference to the key signature and assuming
everything was in a major key. This wasn't true, and this edit
corrects the K fields for those tunes in related keys. Thanks
to John hinton for the data.

This edit required a corresponding change to sorter.py to allow it
to filter on multiple keys and changes to the description of the
various Dulcimer tabs.